# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MarketOverviewConnectionState = typeof MarketOverviewConnectionState[keyof typeof MarketOverviewConnectionState];


export const MarketOverviewConnectionState = {
  connected: 'connected',
  reconnecting: 'reconnecting',
  offline: 'offline',
} as const;

export interface MarketOverview {
  spotPrice: number;
  change24h: number;
  volume24h: number;
  fundingRate: number;
  currencyRate: number;
  lastUpdated: string;
  connectionState: MarketOverviewConnectionState;
}

export interface OptionChainRow {
  id: string;
  strike: number;
  callLtp: number;
  callChange: number;
  callVolume: number;
  putLtp: number;
  putChange: number;
  putVolume: number;
  isAtm: boolean;
}

export type PositionSide = typeof PositionSide[keyof typeof PositionSide];


export const PositionSide = {
  CE: 'CE',
  PE: 'PE',
} as const;

export type PositionStatus = typeof PositionStatus[keyof typeof PositionStatus];


export const PositionStatus = {
  open: 'open',
  'target-hit': 'target-hit',
  'stop-hit': 'stop-hit',
  closed: 'closed',
} as const;

export interface Position {
  id: string;
  instrument: string;
  side: PositionSide;
  entryPrice: number;
  livePrice: number;
  quantity: number;
  pnl: number;
  pnlPercent: number;
  targetPrice: number;
  stopPrice: number;
  status: PositionStatus;
}

export interface TradeActivity {
  id: string;
  type: string;
  instrument: string;
  price: number;
  quantity: number;
  timestamp: string;
}

export interface Portfolio {
  walletBalance: number;
  availableBalance: number;
  totalPnl: number;
  positions: Position[];
  activity: TradeActivity[];
}

export type AdvisorRecommendationDirection = typeof AdvisorRecommendationDirection[keyof typeof AdvisorRecommendationDirection];


export const AdvisorRecommendationDirection = {
  bullish: 'bullish',
  bearish: 'bearish',
  neutral: 'neutral',
} as const;

export interface AdvisorRecommendation {
  id: string;
  title: string;
  body: string;
  instrument: string;
  strike: number;
  direction: AdvisorRecommendationDirection;
  confidence: number;
  createdAt: string;
}

export type PaperOrderInputSide = typeof PaperOrderInputSide[keyof typeof PaperOrderInputSide];


export const PaperOrderInputSide = {
  CE: 'CE',
  PE: 'PE',
} as const;

export interface PaperOrderInput {
  instrument: string;
  side: PaperOrderInputSide;
  entryPrice: number;
  quantity: number;
  targetPercent: number;
  stopPercent: number;
}

export type GetOptionChainParams = {
symbol?: string;
};