# lib/api-zod/src/generated/types/advisorRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdvisorRecommendationDirection } from './advisorRecommendationDirection';

export interface AdvisorRecommendation {
  id: string;
  title: string;
  body: string;
  instrument: string;
  strike: number;
  direction: AdvisorRecommendationDirection;
  confidence: number;
  createdAt: string;
}