# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketOverviewConnectionState } from './marketOverviewConnectionState';

export interface MarketOverview {
  spotPrice: number;
  change24h: number;
  volume24h: number;
  fundingRate: number;
  currencyRate: number;
  lastUpdated: string;
  connectionState: MarketOverviewConnectionState;
}