# lib/api-zod/src/generated/types/optionChainRow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OptionChainRow {
  id: string;
  strike: number;
  callLtp: number;
  callChange: number;
  callVolume: number;
  putLtp: number;
  putChange: number;
  putVolume: number;
  isAtm: boolean;
}