# lib/api-zod/src/generated/types/paperOrderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaperOrderInputSide } from './paperOrderInputSide';

export interface PaperOrderInput {
  instrument: string;
  side: PaperOrderInputSide;
  entryPrice: number;
  quantity: number;
  targetPercent: number;
  stopPercent: number;
}