# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Position } from './position';
import type { TradeActivity } from './tradeActivity';

export interface Portfolio {
  walletBalance: number;
  availableBalance: number;
  totalPnl: number;
  positions: Position[];
  activity: TradeActivity[];
}