# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PositionSide } from './positionSide';
import type { PositionStatus } from './positionStatus';

export interface Position {
  id: string;
  instrument: string;
  side: PositionSide;
  entryPrice: number;
  livePrice: number;
  quantity: number;
  pnl: number;
  pnlPercent: number;
  targetPrice: number;
  stopPrice: number;
  status: PositionStatus;
}